refactor the wasmi_wasi crate
It now mirrors the API of the wasmtime_wasi crate.
