+++ crates/wasi/src/sync/snapshots/mod.rs
